Abstracts image insertion code into separate shared file, so it can be used by all capi templates.
